gateware.platform: update default USB PHY to aux/host
Now that port sharing is not setup automagically, using the control port
PHY by default is not ideal as it needs extra setup to work. This change
switches the default to aux as it's just a standard device port, and
also changes older revisions to use their host PHY to match.